Rework binding system Confirmed/unconfirmed signals Signature caching Fix potential bugs/handle edge cases in SymbolRepository.php::fetchCandles() Define return types Minor refactor

+++ database/migrations/skip/2021_05_25_110329_create_symbols_table.php
@@ -18,7 +18,7 @@ class CreateSymbolsTable extends Migration
             $table->string('symbol', 20)->index();
             $table->string('interval', 3)->index();
             $table->foreignId('exchange_id');
-            $table->integer('last_update', unsigned: true)->default(0);
+            $table->bigInteger('last_update', unsigned: true)->default(0);
             $table->unique(['exchange_id', 'symbol', 'interval']);
             $table->timestamps();
         });
